requirements.txt
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
fake-useragent==1.5.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pillow==11.0.0
py-ulid==1.0.3
pydantic==2.10.1
pydantic_core==2.27.1
pylib-0xe==0.0.5
PySocks==1.7.1
requests==2.32.3
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
tqdm
